pyrogram_client/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.28.0
click==8.1.7
fastapi==0.103.2
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pyaes==1.6.1
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
sqladmin==0.15.1
SQLAlchemy==2.0.21
starlette==0.27.0
TgCrypto==1.2.5
typing_extensions==4.8.0
tzlocal==5.1
uvicorn==0.23.2
WTForms==3.0.1
